Renames Rectangle.{hpp,cpp} to Compositor.{hpp,cpp}
