+++ PlayRho/PlayRho.hpp
@@ -114,7 +114,7 @@ For a more elaborate example, that's of an entire application, see
 #include <PlayRho/Collision/Shapes/EdgeShapeConf.hpp>
 #include <PlayRho/Collision/Shapes/MultiShapeConf.hpp>
 #include <PlayRho/Collision/Shapes/PolygonShapeConf.hpp>
-#include <PlayRho/Collision/Shapes/Rectangle.hpp>
+#include <PlayRho/Collision/Shapes/Compositor.hpp>
 
 // For any and all joint configurations, add one or more of the following.
 #include <PlayRho/Dynamics/Joints/DistanceJointConf.hpp>
